Data types
* Date/Time/Duration datatypes.

* Numeric datatypes.

* Formatting and creating values for Cell.

* Validating constraints.

* Git issues.

+++ src/main/java/com/malyvoj3/csvwvalidator/domain/metadata/properties/IntegerAtomicProperty.java
@@ -1,8 +1,8 @@
 package com.malyvoj3.csvwvalidator.domain.metadata.properties;
 
-public class IntegerAtomicProperty extends AtomicProperty<Integer> {
+public class IntegerAtomicProperty extends AtomicProperty<Long> {
 
-    public IntegerAtomicProperty(Integer value) {
+    public IntegerAtomicProperty(Long value) {
         super(value);
     }
 }
